requirements.txt
Babel==2.3.4
blinker==1.4
coverage==4.1
DateTime==4.1.1
decorator==4.0.9
Flask==0.10.1
Flask-Babel==0.10.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
flipflop==1.0
guess-language==0.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pbr==1.10.0
psycopg2==2.6.1
python-openid==2.2.5
pytz==2016.4
requests==2.10.0
six==1.10.0
speaklater==1.3
SQLAlchemy==1.0.13
sqlalchemy-migrate==0.10.0
sqlparse==0.1.19
Tempita==0.5.2
virtualenv==15.0.1
Werkzeug==0.11.10
Whoosh==2.7.4
WTForms==2.1
zope.interface==4.2.0
